requirements.txt
torch
numpy
Pillow
glob2
statistics
matplotlib